+++ ChangeLog.rst
@@ -4,6 +4,10 @@ Django Rest Framework JWT 2FA Change Log
 Unreleased
 ----------
 
+* Fire ``user_logged_in`` signal on successful authentication
+
+  * Note: This updates ``user.last_login`` by default
+
 * Drop support for Python 3.10, 3.11 and Django 2.2
 
 * Add type annotations to all public and private methods